Subdomain support

+++ app/api.js
@@ -1,49 +1,56 @@
 import app from './express'
 import moment from 'moment-timezone'
+import _ from 'lodash'
 
 import mailself from '../lib/mail'
-import {cache, seed, setup, importJson} from '../data/seed'
+import { cache, seed, setup, importJson } from '../data/seed'
+import { ALLOW_ADMIN_ACTIONS } from '../data/config'
 
-import Promises, {sequelize} from '../models/promise'
+import Promises, { sequelize } from '../models/promise'
+import { Users } from '../models/user'
 import parsePromise from '../lib/parse/promise'
 import parseCredit from '../lib/parse/credit'
 
+const userQuery = (user) => ({
+  model: Users,
+  where: { username: user }
+})
+
 // Actions
 
-app.get('/promises/remove/:id(*)', (req, resp) => {
+app.get('/_s/:user/promises/remove/:id(*)', (req, resp) => {
   console.log('remove', req.params)
   // FIXME: refactor/secure this
   Promises.destroy({
     where: {
       id: req.params.id
-    }
+    },
+    include: [userQuery(req.params.user)],
   }).then(function(deletedRows) {
     console.log('promise removed', deletedRows)
     resp.redirect('/')
   })
 })
 
-app.get('/promises/:user/remove', function(req, resp) {
-  var dbPromises = {}
+app.get('/_s/:user/promises/remove', function(req, resp) {
   Promises.destroy({
-    where: {
-      user: req.params.user
-    }
+    include: [userQuery(req.params.user)],
   }).then(function(deletedRows) {
     console.log('user promises removed', deletedRows)
     resp.redirect('/')
   })
 })
 
-app.get('/promises/complete/:id(*)', (req, resp) => {
+app.get('/_s/:user/promises/complete/:id(*)', (req, resp) => {
   Promises.findOne({
     where: {
       id: req.params.id
-    }
+    },
+    include: [userQuery(req.params.user)],
   }).then(function(promise) {
     promise.update({
       tfin: moment(), //.tz('America/New_York')  FIXME,
-      cred: parseCredit({dueDate: promise.tdue})
+      cred: parseCredit({ dueDate: promise.tdue })
     })
 
     console.log('complete promise', promise.dataValues)
@@ -51,44 +58,51 @@ app.get('/promises/complete/:id(*)', (req, resp) => {
   })
 })
 
-app.post('/promises/edit/:id(*)', (req, res) => {
-  console.log('edit promise', req.params.id, req.body)
+app.post('/_s/:user/promises/edit/:id(*)', (req, res) => {
+  // invalid dates/empty string values should unset db fields
+  const data = _.mapValues(req.body, (value) =>
+    _.includes(['Invalid date', ''], value) ? null : value)
+
+  console.log('edit promise**', req.params.id, data, req.body)
 
   Promises.find({
     where: {
       id: req.params.id
-    }
+    },
+    include: [userQuery(req.params.user)],
   }).then(function(promise) {
     promise.update({
-      cred: parseCredit({dueDate: promise.tdue, finishDate: promise.tfin}),
-      ...req.body
-    })
-
-    console.log('edit promise', promise)
-    if (promise && req.params.id) {
-      res.redirect(`/${req.params.id}`)
-    } else {
-      res.redirect('/')
-    }
-  })
-})
-
-app.get('/promises/create/:urtext(*)', (req, resp) => {
-  console.log('create', req.params)
-  parsePromise({urtext: req.params.urtext, ip: req.ip}).then((parsedPromise) => {
-    Promises.create(parsedPromise).then(function(promise) {
-      console.log('promise created', promise.dataValues)
-      mailself('PROMISE', promise.urtext) // send dreeves@ an email
-      resp.redirect(`/${req.params.urtext}`)
+      cred: parseCredit({ dueDate: promise.tdue, finishDate: promise.tfin }),
+      ...data
+    }).then(function(prom) {
+      console.log('promise updated', req.body)
+      if (promise) {
+        res.redirect(`/${prom.urtext}`)
+      } else {
+        res.redirect('/')
+      }
     })
   })
 })
+
+// Deprecated because we no longer use /create - FIXME: delete?
+//
+// app.get('/promises/create/:urtext(*)', (req, resp) => {
+//   console.log('create', req.params)
+//   parsePromise({urtext: req.params.urtext, ip: req.ip}).then((parsedPromise) => {
+//     Promises.create(parsedPromise).then(function(promise) {
+//       console.log('promise created', promise.dataValues)
+//       mailself('PROMISE', promise.urtext) // send dreeves@ an email
+//       resp.redirect(`/${req.params.urtext}`)
+//     })
+//   })
+// })
 
 // Endpoints
 
 app.get('/promise/:udp/:urtext', function(req, resp) {
   let urtext = req.originalUrl.substr(9)
-  Promises.findOne({where: { urtext }}).then(function(promise) {
+  Promises.findOne({ where: { urtext } }).then(function(promise) {
     console.log('single promise', urtext, promise)
     // resp.write(promise)
     resp.json(promise)
@@ -97,8 +111,8 @@ app.get('/promise/:udp/:urtext', function(req, resp) {
 })
 
 app.get('/promises', function(req, resp) {
-  var dbPromises = {}
-  Promises.findAll({order: sequelize.literal('tdue DESC')}).then(function(promises) {
+  let dbPromises = {}
+  Promises.findAll({ order: sequelize.literal('tdue DESC') }).then(function(promises) {
     // console.log('all promises', promises)
     // create nested array of promises by user:
     promises.forEach(function(promise) {
@@ -110,7 +124,7 @@ app.get('/promises', function(req, resp) {
 })
 
 app.get('/promises/:user', function(req, resp) {
-  var dbPromises = {}
+  let dbPromises = {}
   Promises.findAll({
     where: {
       user: req.params.user
@@ -127,29 +141,30 @@ app.get('/promises/:user', function(req, resp) {
   })
 })
 
-/* Utils */
-
-// drop db and repopulate
-app.get('/import', (req, resp) => {
-  importJson()
-  resp.redirect('/')
-})
-
-// drop db and repopulate
-app.get('/reset', (req, resp) => {
-  seed()
-  setup()
-  resp.redirect('/')
-})
-
 // calculate and store reliability for each user
 app.get('/cache', (req, resp) => {
   cache()
   resp.redirect('/')
 })
 
-// removes all entries from the promises table
-app.get('/empty', (req, resp) => {
-  Promises.destroy({where: {}})
-  resp.redirect('/')
-})
+if (ALLOW_ADMIN_ACTIONS) {
+  /* Utils */
+
+  // insert promises.json into db
+  app.get('/import', (req, resp) => {
+    importJson()
+    resp.redirect('/')
+  })
+
+  // drop db and repopulate
+  app.get('/reset', (req, resp) => {
+    seed()
+    resp.redirect('/')
+  })
+
+  // removes all entries from the promises table
+  app.get('/empty', (req, resp) => {
+    Promises.destroy({ where: {} })
+    resp.redirect('/')
+  })
+}
